C API extensions implement feedback after review
* remove cosmetic static casts in new fuctions
* move attribute helpers into cuopt_c.cpp instead of a new file
* add cuOptGetConstraintMatrixCSR and deprecate cuOptGetConstraintMatrix
* revise/remove comments
