+++ cpp/src/pdlp/cuopt_c.cpp
@@ -21,6 +21,7 @@
 
 #include <cuopt/version_config.hpp>
 
+#include <algorithm>
 #include <cstdlib>
 #include <memory>
 #include <span>
@@ -30,11 +31,13 @@
 using cuopt::mathematical_optimization::char_to_var_type;
 using cuopt::mathematical_optimization::get_memory_backend_type;
 using cuopt::mathematical_optimization::is_valid_public_var_type_code;
+using cuopt::mathematical_optimization::optimization_problem_interface_t;
 using cuopt::mathematical_optimization::problem_and_stream_view_t;
 using cuopt::mathematical_optimization::problem_category_t;
 using cuopt::mathematical_optimization::solution_and_stream_view_t;
 using cuopt::mathematical_optimization::solver_settings_t;
 using cuopt::mathematical_optimization::var_t;
+using cuopt::mathematical_optimization::var_type_to_char;
 using cuopt::mathematical_optimization::io::mps_data_model_t;
 
 class c_get_solution_callback_t : public cuopt::internals::get_solution_callback_t {
@@ -91,6 +94,76 @@ solver_settings_handle_t* get_settings_handle(cuOptSolverSettings settings)
 }
 
 namespace {
+
+// ---- Generic problem-attribute helpers (used by cuOptGetProblem*Attribute below) ----
+
+problem_and_stream_view_t* as_problem(cuOptOptimizationProblem problem)
+{
+  return static_cast<problem_and_stream_view_t*>(problem);
+}
+
+optimization_problem_interface_t<cuopt_int_t, cuopt_float_t>* get_iface(
+  cuOptOptimizationProblem problem)
+{
+  return as_problem(problem)->get_problem();
+}
+
+bool is_int_attribute(cuopt_int_t attribute)
+{
+  switch (attribute) {
+    case CUOPT_ATTR_NUM_VARIABLES:
+    case CUOPT_ATTR_NUM_CONSTRAINTS:
+    case CUOPT_ATTR_NUM_NONZEROS:
+    case CUOPT_ATTR_NUM_INTEGERS:
+    case CUOPT_ATTR_OBJECTIVE_SENSE:
+    case CUOPT_ATTR_PROBLEM_CATEGORY:
+    case CUOPT_ATTR_IS_MIP:
+    case CUOPT_ATTR_HAS_QUADRATIC_OBJECTIVE:
+    case CUOPT_ATTR_HAS_QUADRATIC_CONSTRAINTS: return true;
+    default: return false;
+  }
+}
+
+bool is_float_attribute(cuopt_int_t attribute)
+{
+  return attribute == CUOPT_ATTR_OBJECTIVE_OFFSET ||
+         attribute == CUOPT_ATTR_OBJECTIVE_SCALING_FACTOR;
+}
+
+bool is_float_array_attribute(cuopt_int_t attribute)
+{
+  switch (attribute) {
+    case CUOPT_ARRAY_ATTR_OBJECTIVE_COEFFICIENTS:
+    case CUOPT_ARRAY_ATTR_VARIABLE_LOWER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_VARIABLE_UPPER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_LOWER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_UPPER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_RHS: return true;
+    default: return false;
+  }
+}
+
+bool is_char_array_attribute(cuopt_int_t attribute)
+{
+  return attribute == CUOPT_ARRAY_ATTR_CONSTRAINT_SENSE ||
+         attribute == CUOPT_ARRAY_ATTR_VARIABLE_TYPES;
+}
+
+cuopt_int_t get_array_size(optimization_problem_interface_t<cuopt_int_t, cuopt_float_t>* problem,
+                           cuopt_int_t attribute)
+{
+  switch (attribute) {
+    case CUOPT_ARRAY_ATTR_OBJECTIVE_COEFFICIENTS:
+    case CUOPT_ARRAY_ATTR_VARIABLE_LOWER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_VARIABLE_UPPER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_VARIABLE_TYPES: return problem->get_n_variables();
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_LOWER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_UPPER_BOUNDS:
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_RHS:
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_SENSE: return problem->get_n_constraints();
+    default: return -1;
+  }
+}
 
 void coo_to_csr(cuopt_int_t num_entries,
                 const cuopt_int_t* row_index,
@@ -168,6 +241,10 @@ constexpr char k_deprecated_quadratic_ranged_problem_msg[] =
   "cuOptCreateQuadraticRangedProblem is deprecated. Use cuOptCreateRangedProblem to set up the "
   "linear problem, then cuOptSetQuadraticObjective to specify the quadratic objective terms. "
   "For QCQP models, call cuOptAddQuadraticConstraint for each quadratic constraint.";
+
+constexpr char k_deprecated_get_constraint_matrix_msg[] =
+  "cuOptGetConstraintMatrix is deprecated. Use cuOptGetConstraintMatrixCSR for identical CSR "
+  "output, or cuOptGetConstraintMatrixCSC for compressed sparse column format.";
 
 }  // namespace
 
@@ -715,25 +792,11 @@ cuopt_int_t cuOptGetConstraintMatrix(cuOptOptimizationProblem problem,
                                      cuopt_int_t* constraint_matrix_column_indices_ptr,
                                      cuopt_float_t* constraint_matrix_coefficients_ptr)
 {
-  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
-  if (constraint_matrix_row_offsets_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
-  if (constraint_matrix_column_indices_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
-  if (constraint_matrix_coefficients_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
-  problem_and_stream_view_t* problem_and_stream_view =
-    static_cast<problem_and_stream_view_t*>(problem);
-
-  auto* prob           = problem_and_stream_view->get_problem();
-  cuopt_int_t num_nnz  = prob->get_nnz();
-  cuopt_int_t num_rows = prob->get_n_constraints();
-
-  prob->copy_constraint_matrix_to_host(constraint_matrix_coefficients_ptr,
-                                       constraint_matrix_column_indices_ptr,
-                                       constraint_matrix_row_offsets_ptr,
-                                       num_nnz,
-                                       num_nnz,
-                                       num_rows + 1);
-
-  return CUOPT_SUCCESS;
+  CUOPT_LOG_WARN("%s", k_deprecated_get_constraint_matrix_msg);
+  return cuOptGetConstraintMatrixCSR(problem,
+                                     constraint_matrix_row_offsets_ptr,
+                                     constraint_matrix_column_indices_ptr,
+                                     constraint_matrix_coefficients_ptr);
 }
 
 cuopt_int_t cuOptGetConstraintSense(cuOptOptimizationProblem problem, char* constraint_sense_ptr)
@@ -1294,4 +1357,227 @@ cuopt_int_t cuOptGetReducedCosts(cuOptSolution solution, cuopt_float_t* reduced_
   } catch (const std::logic_error&) {
     return CUOPT_INVALID_ARGUMENT;
   }
+}
+
+/* -------------------------------------------------------------------------- */
+/* Generic problem attribute getters (declared in cuopt_c.h)                  */
+/* -------------------------------------------------------------------------- */
+
+cuopt_int_t cuOptGetProblemIntAttribute(cuOptOptimizationProblem problem,
+                                        cuopt_int_t attribute,
+                                        cuopt_int_t* value_out)
+{
+  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (value_out == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (!is_int_attribute(attribute)) { return CUOPT_INVALID_ARGUMENT; }
+
+  auto* iface = get_iface(problem);
+  switch (attribute) {
+    case CUOPT_ATTR_NUM_VARIABLES: *value_out = iface->get_n_variables(); return CUOPT_SUCCESS;
+    case CUOPT_ATTR_NUM_CONSTRAINTS: *value_out = iface->get_n_constraints(); return CUOPT_SUCCESS;
+    case CUOPT_ATTR_NUM_NONZEROS: *value_out = iface->get_nnz(); return CUOPT_SUCCESS;
+    case CUOPT_ATTR_NUM_INTEGERS: *value_out = iface->get_n_integers(); return CUOPT_SUCCESS;
+    case CUOPT_ATTR_OBJECTIVE_SENSE:
+      *value_out = iface->get_sense() ? CUOPT_MAXIMIZE : CUOPT_MINIMIZE;
+      return CUOPT_SUCCESS;
+    case CUOPT_ATTR_PROBLEM_CATEGORY:
+      *value_out = static_cast<cuopt_int_t>(iface->get_problem_category());
+      return CUOPT_SUCCESS;
+    case CUOPT_ATTR_IS_MIP: {
+      const auto category = iface->get_problem_category();
+      *value_out =
+        (category == problem_category_t::MIP || category == problem_category_t::IP) ? 1 : 0;
+      return CUOPT_SUCCESS;
+    }
+    case CUOPT_ATTR_HAS_QUADRATIC_OBJECTIVE:
+      *value_out = iface->has_quadratic_objective() ? 1 : 0;
+      return CUOPT_SUCCESS;
+    case CUOPT_ATTR_HAS_QUADRATIC_CONSTRAINTS:
+      *value_out = iface->has_quadratic_constraints() ? 1 : 0;
+      return CUOPT_SUCCESS;
+    default: return CUOPT_INVALID_ARGUMENT;
+  }
+}
+
+cuopt_int_t cuOptGetProblemFloatAttribute(cuOptOptimizationProblem problem,
+                                          cuopt_int_t attribute,
+                                          cuopt_float_t* value_out)
+{
+  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (value_out == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (!is_float_attribute(attribute)) { return CUOPT_INVALID_ARGUMENT; }
+
+  auto* iface = get_iface(problem);
+  if (attribute == CUOPT_ATTR_OBJECTIVE_OFFSET) {
+    *value_out = iface->get_objective_offset();
+  } else {
+    *value_out = iface->get_objective_scaling_factor();
+  }
+  return CUOPT_SUCCESS;
+}
+
+cuopt_int_t cuOptGetProblemFloatArrayAttribute(cuOptOptimizationProblem problem,
+                                               cuopt_int_t attribute,
+                                               cuopt_float_t* out,
+                                               cuopt_int_t count)
+{
+  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (out == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (!is_float_array_attribute(attribute)) { return CUOPT_INVALID_ARGUMENT; }
+
+  auto* iface                = get_iface(problem);
+  const cuopt_int_t expected = get_array_size(iface, attribute);
+  if (expected < 0 || count != expected) { return CUOPT_INVALID_ARGUMENT; }
+
+  std::vector<cuopt_float_t> values;
+  switch (attribute) {
+    case CUOPT_ARRAY_ATTR_OBJECTIVE_COEFFICIENTS:
+      values = iface->get_objective_coefficients_host();
+      break;
+    case CUOPT_ARRAY_ATTR_VARIABLE_LOWER_BOUNDS:
+      values = iface->get_variable_lower_bounds_host();
+      break;
+    case CUOPT_ARRAY_ATTR_VARIABLE_UPPER_BOUNDS:
+      values = iface->get_variable_upper_bounds_host();
+      break;
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_LOWER_BOUNDS:
+      values = iface->get_constraint_lower_bounds_host();
+      break;
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_UPPER_BOUNDS:
+      values = iface->get_constraint_upper_bounds_host();
+      break;
+    case CUOPT_ARRAY_ATTR_CONSTRAINT_RHS: values = iface->get_constraint_bounds_host(); break;
+    default: return CUOPT_INVALID_ARGUMENT;
+  }
+
+  if (values.size() != expected) { return CUOPT_VALIDATION_ERROR; }
+  std::copy(values.begin(), values.end(), out);
+  return CUOPT_SUCCESS;
+}
+
+cuopt_int_t cuOptGetProblemCharArrayAttribute(cuOptOptimizationProblem problem,
+                                              cuopt_int_t attribute,
+                                              char* out,
+                                              cuopt_int_t count)
+{
+  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (out == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (!is_char_array_attribute(attribute)) { return CUOPT_INVALID_ARGUMENT; }
+
+  auto* iface                = get_iface(problem);
+  const cuopt_int_t expected = get_array_size(iface, attribute);
+  if (expected < 0 || count != expected) { return CUOPT_INVALID_ARGUMENT; }
+
+  if (attribute == CUOPT_ARRAY_ATTR_CONSTRAINT_SENSE) {
+    const std::vector<char> row_types = iface->get_row_types_host();
+    if (row_types.size() != expected) { return CUOPT_VALIDATION_ERROR; }
+    std::copy(row_types.begin(), row_types.end(), out);
+  } else if (attribute == CUOPT_ARRAY_ATTR_VARIABLE_TYPES) {
+    const std::vector<var_t> var_types = iface->get_variable_types_host();
+    if (var_types.size() != expected) { return CUOPT_VALIDATION_ERROR; }
+    for (cuopt_int_t i = 0; i < count; ++i) {
+      out[i] = var_type_to_char(var_types[i]);
+    }
+  } else {
+    return CUOPT_INVALID_ARGUMENT;
+  }
+  return CUOPT_SUCCESS;
+}
+
+cuopt_int_t cuOptGetProblemStringArrayAttribute(cuOptOptimizationProblem problem,
+                                                cuopt_int_t attribute,
+                                                const char** strings_out,
+                                                cuopt_int_t count)
+{
+  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (strings_out == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (attribute != CUOPT_STRING_ARRAY_VARIABLE_NAMES && attribute != CUOPT_STRING_ARRAY_ROW_NAMES) {
+    return CUOPT_INVALID_ARGUMENT;
+  }
+
+  auto* iface       = get_iface(problem);
+  const auto& names = (attribute == CUOPT_STRING_ARRAY_VARIABLE_NAMES) ? iface->get_variable_names()
+                                                                       : iface->get_row_names();
+
+  if (names.size() != count) { return CUOPT_INVALID_ARGUMENT; }
+  for (cuopt_int_t i = 0; i < count; ++i) {
+    strings_out[i] = names[i].c_str();
+  }
+  return CUOPT_SUCCESS;
+}
+
+cuopt_int_t cuOptGetConstraintMatrixCSR(cuOptOptimizationProblem problem,
+                                        cuopt_int_t* constraint_matrix_row_offsets_ptr,
+                                        cuopt_int_t* constraint_matrix_column_indices_ptr,
+                                        cuopt_float_t* constraint_matrix_coefficients_ptr)
+{
+  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (constraint_matrix_row_offsets_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (constraint_matrix_column_indices_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (constraint_matrix_coefficients_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+
+  auto* iface          = get_iface(problem);
+  cuopt_int_t num_nnz  = iface->get_nnz();
+  cuopt_int_t num_rows = iface->get_n_constraints();
+
+  iface->copy_constraint_matrix_to_host(constraint_matrix_coefficients_ptr,
+                                        constraint_matrix_column_indices_ptr,
+                                        constraint_matrix_row_offsets_ptr,
+                                        num_nnz,
+                                        num_nnz,
+                                        num_rows + 1);
+  return CUOPT_SUCCESS;
+}
+
+cuopt_int_t cuOptGetConstraintMatrixCSC(cuOptOptimizationProblem problem,
+                                        cuopt_int_t* column_offsets_ptr,
+                                        cuopt_int_t* row_indices_ptr,
+                                        cuopt_float_t* values_ptr)
+{
+  if (problem == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+  if (column_offsets_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+
+  auto* iface         = get_iface(problem);
+  const cuopt_int_t n = iface->get_n_variables();
+  const cuopt_int_t m = iface->get_n_constraints();
+
+  const std::vector<cuopt_int_t> row_offsets = iface->get_constraint_matrix_offsets_host();
+  const std::vector<cuopt_int_t> col_indices = iface->get_constraint_matrix_indices_host();
+  const std::vector<cuopt_float_t> values    = iface->get_constraint_matrix_values_host();
+  const cuopt_int_t nnz                      = static_cast<cuopt_int_t>(values.size());
+
+  // Empty / unset matrix: emit all-zero column offsets and nothing else.
+  if (row_offsets.size() < m + 1 || nnz == 0) {
+    for (cuopt_int_t c = 0; c <= n; ++c) {
+      column_offsets_ptr[c] = 0;
+    }
+    return CUOPT_SUCCESS;
+  }
+  if (row_indices_ptr == nullptr || values_ptr == nullptr) { return CUOPT_INVALID_ARGUMENT; }
+
+  // Count non-zeros per column, then prefix-sum into column offsets.
+  std::vector<cuopt_int_t> col_counts(n, 0);
+  for (cuopt_int_t k = 0; k < nnz; ++k) {
+    const cuopt_int_t c = col_indices[k];
+    if (c < 0 || c >= n) { return CUOPT_VALIDATION_ERROR; }
+    ++col_counts[c];
+  }
+  column_offsets_ptr[0] = 0;
+  for (cuopt_int_t c = 0; c < n; ++c) {
+    column_offsets_ptr[c + 1] = column_offsets_ptr[c] + col_counts[c];
+  }
+
+  // Scatter each CSR entry into its CSC position using a running write cursor per column.
+  std::vector<cuopt_int_t> next(column_offsets_ptr, column_offsets_ptr + n);
+  for (cuopt_int_t i = 0; i < m; ++i) {
+    const cuopt_int_t row_begin = row_offsets[i];
+    const cuopt_int_t row_end   = row_offsets[i + 1];
+    for (cuopt_int_t k = row_begin; k < row_end; ++k) {
+      const cuopt_int_t c    = col_indices[k];
+      const cuopt_int_t dest = next[c]++;
+      row_indices_ptr[dest]  = i;
+      values_ptr[dest]       = values[k];
+    }
+  }
+  return CUOPT_SUCCESS;
 }
